feat: embed cover art
