+++ src/web/sample.env
@@ -51,6 +51,8 @@ LIBRARY_NAME=
 YOUTUBE_API_KEY=
 # Custom file extension for tracks (e.g mp3) (default: opus)
 # TRACK_EXTENSION=opus
+# Include cover art in downloaded files (default: false)
+# EMBED_COVER_ART=false
 # Custom path to ffmpeg binary (default: defined in $PATH)
 # FFMPEG_PATH=
 # Custom path to yt-dlp binary (default: defined in $PATH)
